Add missing int casts in pool model

+++ pypolygames/model_zoo/res_conv_conv_logit_pool_model.py
@@ -94,7 +94,7 @@ class ResConvConvLogitPoolModel(torch.jit.ScriptModule):
         mono = [
             nn.Conv2d(
                 c,
-                nnsize * c,
+                int(nnsize * c),
                 nnks,
                 stride=stride,
                 padding=padding,
@@ -107,8 +107,8 @@ class ResConvConvLogitPoolModel(torch.jit.ScriptModule):
         for i in range(nb_nets):
             nets = [
                 nn.Conv2d(
-                    nnsize * c,
-                    nnsize * c,
+                    int(nnsize * c),
+                    int(nnsize * c),
                     nnks,
                     stride=stride,
                     padding=padding,
@@ -122,7 +122,7 @@ class ResConvConvLogitPoolModel(torch.jit.ScriptModule):
                     nets[j] = nn.Sequential(
                         nets[j],
                         nn.BatchNorm2d(
-                            nnsize * c, track_running_stats=True, affine=bn_affine
+                            int(nnsize * c), track_running_stats=True, affine=bn_affine
                         ),
                     )
             if pooling:
@@ -139,23 +139,23 @@ class ResConvConvLogitPoolModel(torch.jit.ScriptModule):
             resnet_list.append(nets)
         if bn or bn_affine:
             mono.append(
-                nn.BatchNorm2d(nnsize * c, track_running_stats=True, affine=bn_affine),
+                nn.BatchNorm2d(int(nnsize * c), track_running_stats=True, affine=bn_affine),
             )
             for i in range(nb_nets):
                 for j in range(nb_layers_per_net):
                     resnet_list[i][j] = nn.Sequential(
                         resnet_list[i][j],
                         nn.BatchNorm2d(
-                            nnsize * c, track_running_stats=True, affine=bn_affine
+                            int(nnsize * c), track_running_stats=True, affine=bn_affine
                         ),
                     )
         for i in range(nb_nets):
             resnet_list[i] = nn.ModuleList(resnet_list[i])
         self.mono = nn.Sequential(*mono)
         self.resnets = nn.ModuleList(resnet_list)
-        self.v = nn.Linear(2 * nnsize * c, 1)
+        self.v = nn.Linear(2 * int(nnsize * c), 1)
         self.pi_logit = nn.Conv2d(
-            nnsize * c, c_prime, nnks, stride=stride, padding=padding, dilation=dilation
+            int(nnsize * c), c_prime, nnks, stride=stride, padding=padding, dilation=dilation
         )
 
     @torch.jit.script_method
